# lib/api-zod/src/generated/types/aiConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UNS | أُنس — API specification
 * OpenAPI spec version: 0.2.0
 */
import type { AiConfigCrisisThreshold } from "./aiConfigCrisisThreshold";
import type { AiConfigDefaultDialect } from "./aiConfigDefaultDialect";
import type { AiConfigModelTier } from "./aiConfigModelTier";
import type { AiConfigToneIntensity } from "./aiConfigToneIntensity";

export interface AiConfig {
  defaultDialect: AiConfigDefaultDialect;
  toneIntensity: AiConfigToneIntensity;
  spiritualLayerEnabled: boolean;
  crisisThreshold: AiConfigCrisisThreshold;
  familyModeEnabled: boolean;
  modelTier: AiConfigModelTier;
  systemPromptSuffix?: string;
}