# lib/api-zod/src/generated/types/companionChatResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UNS | أُنس — API specification
 * OpenAPI spec version: 0.2.0
 */
import type { CrisisResource } from "./crisisResource";

export interface CompanionChatResponse {
  response: string;
  emotion?: string;
  crisisDetected: boolean;
  crisisResources?: CrisisResource[];
  messageId: string;
}