# lib/api-zod/src/generated/types/conversationMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UNS | أُنس — API specification
 * OpenAPI spec version: 0.2.0
 */
import type { ConversationMessageRole } from "./conversationMessageRole";

export interface ConversationMessage {
  id: string;
  role: ConversationMessageRole;
  content: string;
  createdAt: Date;
  emotion?: string;
}