# lib/api-zod/src/generated/types/crisisResource.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UNS | أُنس — API specification
 * OpenAPI spec version: 0.2.0
 */

export interface CrisisResource {
  country: string;
  name: string;
  number: string;
}