# lib/api-zod/src/generated/types/moodHistoryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UNS | أُنس — API specification
 * OpenAPI spec version: 0.2.0
 */
import type { MoodEntry } from "./moodEntry";
import type { MoodHistoryResponseSummary } from "./moodHistoryResponseSummary";

export interface MoodHistoryResponse {
  entries: MoodEntry[];
  summary: MoodHistoryResponseSummary;
}