# lib/api-zod/src/generated/types/safetyOverview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UNS | أُنس — API specification
 * OpenAPI spec version: 0.2.0
 */
import type { SafetyEvent } from "./safetyEvent";
import type { SafetyOverviewRegionBreakdownItem } from "./safetyOverviewRegionBreakdownItem";

export interface SafetyOverview {
  eventsThisWeek: number;
  eventsThisMonth: number;
  crisisResponseRate: number;
  regionBreakdown: SafetyOverviewRegionBreakdownItem[];
  recentEvents: SafetyEvent[];
}